Consolidate test parsing into single report_failed_tests.py
- Combine parse_test_output.py into report_failed_tests.py
- Keep unrelated test_rocwmma.py and test_rocprofiler_compute.py changes
- Simplify codebase by having single file for test output parsing
