+++ build_tools/github_actions/test_executable_scripts/test_rocprofiler_compute.py
@@ -17,8 +17,6 @@ ROCPROFILER_COMPUTE_DIRECTORY = THEROCK_PATH / "libexec" / "rocprofiler-compute"
 # Set up excluded tests
 EXCLUDED_TESTS = [
     "test_profile_live_attach_detach",
-    "test_metric_validation",
-    "test_L1_cache_counters",
 ]
 
 # quick Tests
@@ -27,8 +25,10 @@ QUICK_TESTS = [
     "test_utils",
     "test_num_xcds_cli_output",
     "test_num_xcds_spec_class",
+    "test_L1_cache_counters",
     "test_analyze_workloads",
     "test_analyze_commands",
+    "test_metric_validation",
     "test_profile_iteration_multiplexing_1",
 ]
